+++ plugins/gltf-transform/vite.config.js
@@ -1,15 +1,12 @@
 import {defineConfig} from 'vite'
-import json from '@rollup/plugin-json';
 import dts from 'vite-plugin-dts'
 import packageJson from './package.json';
-import license from 'rollup-plugin-license';
-import replace from '@rollup/plugin-replace';
-import glsl from 'rollup-plugin-glsl';
 import path from 'node:path';
+import {commonPlugins, globalsReplacePlugin} from '../../scripts/vite-utils.mjs';
 
 const isProd = process.env.NODE_ENV === 'production'
-const { name, version, author } = packageJson
-const {main, module, browser} = packageJson
+const { name } = packageJson
+const {main, module} = packageJson
 
 const globals = {
     'three': 'threepipe', // just incase someone uses three
@@ -21,9 +18,6 @@ export default defineConfig({
         exclude: ['uiconfig.js', 'ts-browser-helpers'],
     },
     base: '',
-    // define: {
-    //     'process.env': process.env
-    // },
     build: {
         sourcemap: true,
         minify: false,
@@ -45,47 +39,14 @@ export default defineConfig({
         },
         rollupOptions: {
             output: {
-                // inlineDynamicImports: false,
                 globals,
             },
             external: Object.keys(globals),
-
         },
     },
     plugins: [
         isProd ? dts({tsconfigPath: './tsconfig.json'}) : null,
-        replace({
-            'from \'three\'': 'from \'threepipe\'',
-            delimiters: ['', ''],
-            preventAssignment: true,
-        }),
-        replace({
-            'process.env.NODE_ENV': JSON.stringify(isProd ? 'production' : 'development'),
-            preventAssignment: true,
-        }),
-        glsl({ // todo: minify glsl.
-            include: 'src/**/*.glsl',
-        }),
-        json(),
-        // postcss({
-        //     modules: false,
-        //     autoModules: true,  // todo; issues with typescript import css, because inject is false
-        //     inject: false,
-        //     minimize: isProduction,
-        //     // Or with custom options for `postcss-modules`
-        // }),
-        license({
-            banner: `
-        @license
-        ${name} v${version}
-        Copyright 2022<%= moment().format('YYYY') > 2022 ? '-' + moment().format('YYYY') : null %> ${author}
-        ${packageJson.license} License
-        See ./dependencies.txt for any bundled third-party dependencies and licenses.
-      `,
-            thirdParty: {
-                output: path.join(__dirname, 'dist', 'dependencies.txt'),
-                includePrivate: true, // Default is false.
-            },
-        }),
+        ...globalsReplacePlugin(globals, isProd),
+        ...commonPlugins(packageJson, path.resolve(import.meta.dirname || __dirname), isProd),
     ],
 })